+++ editors/zed/languages/ppphp/overrides.scm
@@ -1,0 +1,2 @@
+[(string) (encapsed_string) (heredoc) (nowdoc)] @string
+(comment) @comment.inclusive